Move tests into appropriate modules
